CI模块添加完整的操作审计

+++ CI/lib/api/favorites.pm
@@ -21,6 +21,9 @@ post '/favorites/:groupid' => sub {
     my $user = $api::sso->run( cookie => cookie( $api::cookiekey ), map{ $_ => request->headers->{$_} }qw( appkey appname ));
     return +{ stat => $JSON::false, code => 10000 } unless $user;
 
+    eval{ $api::auditlog->run( user => $user, title => 'ADD FAVORITES', content => "FLOWLINEID:$param->{ciid} NAME:$param->{name}" ); };
+    return +{ stat => $JSON::false, info => $@ } if $@;
+
     eval{ 
         $api::mysql->execute( "replace into favorites (`ciid`,`name`, `user` ) values( '$param->{ciid}', '$param->{name}', '$user' )");
     };
@@ -39,6 +42,9 @@ del '/favorites/:groupid' => sub {
 
     my $user = $api::sso->run( cookie => cookie( $api::cookiekey ), map{ $_ => request->headers->{$_} }qw( appkey appname ));
     return +{ stat => $JSON::false, code => 10000 } unless $user;
+
+    eval{ $api::auditlog->run( user => $user, title => 'DEL FAVORITES', content => "FLOWLINEID:$param->{ciid}" ); };
+    return +{ stat => $JSON::false, info => $@ } if $@;
 
     eval{ 
         $api::mysql->execute( "delete from favorites where ciid='$param->{ciid}' and user='$user'" );
